tbones-funfactAPI/app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
jinja2
itsdangerous
python-multipart